backend/tasks/celery_app.py
"""
celery_app.py — Celery configuration and app instance.
"""

import os
from celery import Celery
from dotenv import load_dotenv

load_dotenv()

REDIS_URL = os.environ.get("REDIS_URL", "redis://localhost:6379/0")

celery = Celery(
    "itifaq_onboarding",
    broker=REDIS_URL,
    backend=REDIS_URL,
    include=[
        "tasks.process_docs",
        "tasks.notifications",
    ],
)

celery.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
    timezone="Asia/Dubai",
    enable_utc=True,
    task_track_started=True,
    worker_prefetch_multiplier=1,
    task_acks_late=True,
)
