requirements.txt
# Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Werkzeug==3.0.3

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
pgvector==0.3.2
alembic==1.13.2

# Background Jobs
celery==5.4.0
redis==5.0.8
flower==2.0.1

# WhatsApp / Twilio
twilio==9.2.3

# AI / OpenAI
openai==1.40.0

# OCR
paddlepaddle==2.6.1
paddleocr==2.7.3
Pillow==10.4.0

# Email
sendgrid==6.11.0

# PDF Generation
reportlab==4.2.2
PyPDF2==3.0.1

# File handling
python-magic==0.4.27
python-dotenv==1.0.1

# Scheduling
requests==2.32.3

# Node interaction (real-time notifications via subprocess or socket)
flask-socketio==5.3.6
eventlet==0.36.1

# Utilities
python-dateutil==2.9.0
shortuuid==1.0.13
cryptography==42.0.8

# Testing
pytest==8.3.2
pytest-flask==1.3.0
coverage==7.6.1
